Support find for restaurant like, update asyncjob

+++ simple_api/modules/restaurantlike/restaurantlikebiz/user_unlike_restaurant.go
@@ -3,11 +3,13 @@ package restaurantlikebiz
 import (
 	"context"
 	"learn-go/simple_api/common"
+	"learn-go/simple_api/component/asyncjob"
 	"learn-go/simple_api/modules/restaurantlike/restaurantlikemodel"
 )
 
 type UserUnlikeRestaurantStore interface {
 	Delete(ctx context.Context, userId, restaurantId int) error
+	Find(ctx context.Context, condition map[string]interface{}) (*restaurantlikemodel.Like, error)
 }
 
 type DecreaseLikeCountStore interface {
@@ -29,6 +31,9 @@ func (biz *userUnlikeRestaurantBiz) UserUnlikeRestaurant(
 ) error {
 	// NOTE: nên find coi coi user đã like restaurant chưa trước khi create
 	// vì tìm theo 2 khoá chính user_id và restaurant_id nên sẽ rất nhanh => nên làm
+	if like, err := biz.store.Find(ctx, map[string]interface{}{"restaurant_id": restaurantId, "user_id": userId}); like == nil {
+		return restaurantlikemodel.ErrCannotUnlikeRestaurant(err)
+	}
 
 	err := biz.store.Delete(ctx, userId, restaurantId)
 
@@ -39,8 +44,18 @@ func (biz *userUnlikeRestaurantBiz) UserUnlikeRestaurant(
 	// This is consider as side effect so no need to handle error
 	go func() {
 		defer common.AppRecover()
-		_ = biz.decreaseLikeStore.DecreaseLikeCount(ctx, restaurantId)
+		job := asyncjob.NewJob(func(ctx context.Context) error {
+			return biz.decreaseLikeStore.DecreaseLikeCount(ctx, restaurantId)
+		})
+
+		_ = asyncjob.NewGroup(true, job).Run(ctx)
 	}()
+
+	// This is consider as side effect so no need to handle error
+	// go func() {
+	// 	defer common.AppRecover()
+	// 	_ = biz.decreaseLikeStore.DecreaseLikeCount(ctx, restaurantId)
+	// }()
 
 	return nil
 }
